lots of stuff

+++ .devcontainer/features/nvidia/install.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 set -e
 
-curl -fsSL https://nvidia.github.io/libnvidia-container/gpgkey | sudo gpg --dearmor -o /usr/share/keyrings/nvidia-container-toolkit-keyring.gpg
-curl -s -L https://nvidia.github.io/libnvidia-container/stable/deb/nvidia-container-toolkit.list |
-sed 's#deb https://#deb [signed-by=/usr/share/keyrings/nvidia-container-toolkit-keyring.gpg] https://#g' |
-	sudo tee /etc/apt/sources.list.d/nvidia-container-toolkit.list
+if [ "${GPU}" = "true" ]; then
+	curl -fsSL https://nvidia.github.io/libnvidia-container/gpgkey | sudo gpg --dearmor -o /usr/share/keyrings/nvidia-container-toolkit-keyring.gpg
+	curl -s -L https://nvidia.github.io/libnvidia-container/stable/deb/nvidia-container-toolkit.list |
+	sed 's#deb https://#deb [signed-by=/usr/share/keyrings/nvidia-container-toolkit-keyring.gpg] https://#g' |
+		sudo tee /etc/apt/sources.list.d/nvidia-container-toolkit.list
 
-sudo apt-get update
-sudo apt-get install -y nvidia-container-toolkit
-rm -rf /var/lib/apt/lists/*
+	sudo apt-get update
+	sudo apt-get install -y nvidia-container-toolkit
+	rm -rf /var/lib/apt/lists/*
+fi
